backend/requirements.txt
flask
flask-pymongo
flask-cors
flask-bcrypt
PyJWT
python-dotenv
